+++ services/madoc-ts/src/frontend/shared/capture-models/new/components/DefaultSingleEntity.tsx
@@ -7,6 +7,8 @@ import { useEntityDetails } from '../hooks/use-entity-details';
 import { useHighlightSelector } from '../hooks/use-highlight-selector';
 import { mapProperties } from '../utility/map-properties';
 import { EditorRenderingConfig, useProfileOverride, useSlotContext } from './EditorSlots';
+import { FieldSet } from '../../../form/FieldSet';
+import { useResolvedSelector } from '../hooks/use-resolved-selector';
 
 export const DefaultSingleEntity: EditorRenderingConfig['SingleEntity'] = props => {
   const { showTitle } = props;
@@ -15,8 +17,9 @@ export const DefaultSingleEntity: EditorRenderingConfig['SingleEntity'] = props 
   const { isModified } = useEntityDetails(entity);
   const entityLabel = getEntityLabel(entity);
   const ProfileSpecificComponent = useProfileOverride('SingleEntity');
+  const [selector, { isBlockingForm: disableForm }] = useResolvedSelector(entity);
 
-  useHighlightSelector(entity.selector?.id);
+  useHighlightSelector(selector?.id);
 
   if (ProfileSpecificComponent) {
     return <ProfileSpecificComponent {...props} />;
@@ -30,19 +33,21 @@ export const DefaultSingleEntity: EditorRenderingConfig['SingleEntity'] = props 
       <Slots.AdjacentNavigation>
         {isModified && <ModifiedStatus />}
         <Slots.InlineSelector />
-        {mapProperties(entity, ({ type, hasSelector, label, description, property, canInlineField }) => {
-          return (
-            <Slots.InlineProperties
-              type={type}
-              property={property}
-              label={label}
-              description={description}
-              canInlineField={canInlineField}
-              hasSelector={hasSelector}
-              disableRemoving
-            />
-          );
-        })}
+        <FieldSet disabled={disableForm}>
+          {mapProperties(entity, ({ type, hasSelector, label, description, property, canInlineField }) => {
+            return (
+              <Slots.InlineProperties
+                type={type}
+                property={property}
+                label={label}
+                description={description}
+                canInlineField={canInlineField}
+                hasSelector={hasSelector}
+                disableRemoving
+              />
+            );
+          })}
+        </FieldSet>
       </Slots.AdjacentNavigation>
     </React.Fragment>
   );